Protect against session fixation attacks
Sometimes, a malicious user will tamper the session
cookie to inject a controlled session id.
Without this protection, the session handler tries to read a
session value with this session id and, if the id is not
found, a new session entry is created using the malicious id.
To avoid that the PHP team introduced the use_strict_mode option
in the native sessions module.
This commit introduces the use of the use_strict_mode option
in the SessionHandler to prevent session fixation by issuing
a new session id for the current session when an unknown session
id is used.
For the sake of facility, we rely only on the opened sessions
to conclude whether or not the session id is known.
We could add opened session ids history as in https://github.com/1ma/RedisSessionHandler
but I can't see any useful use case.